Rewrite cookiecutter tests as pytest tests and fix tests
